re-purpose to delete dynamo item on image deletion.
