+++ metadata-editor/app/lib/MetadataMessageConsumer.scala
@@ -1,35 +1,20 @@
 package lib
 
-import play.api.libs.json._
 import com.gu.mediaservice.lib.aws.MessageConsumer
-import com.gu.mediaservice.model._
-import play.api.libs.concurrent.Execution.Implicits._
+import play.api.libs.json._
 
+import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.Future
 
 object MetadataMessageConsumer extends MessageConsumer(
-  Config.queueUrl, Config.awsEndpoint, Config.awsCredentials, MetadataEditorMetrics.processingLatency) {
+  Config.queueUrl, Config.awsEndpoint, Config.awsCredentials, MetadataEditorMetrics.processingLatency) with DynamoEdits {
 
   override def chooseProcessor(subject: String): Option[JsValue => Future[Any]] =
     PartialFunction.condOpt(subject) {
-      case "image-indexed" => processIndexedImage
+      case "image-deleted" => processDeletedImage
     }
 
-  def isArchivingUsageRights[T <: UsageRights](usageRights: T) =
-    usageRights match {
-      case _:StaffPhotographer | _:ContractPhotographer | _:CommissionedPhotographer => true
-      case _ => false
-    }
-
-  def processIndexedImage(message: JsValue): Future[Any] = {
-    message.asOpt[Image] match {
-      case Some(image: Image) => {
-        isArchivingUsageRights(image.usageRights) match {
-          case true => DynamoEdits.setArchived(image.id, true)
-          case _    => Future(image)
-        }
-      }
-      case _ => sys.error("Couldn't parse json to Image")
-    }
+  def processDeletedImage(message: JsValue) = Future {
+      withImageId(message)(id => dynamo.deleteItem(id))
   }
 }
